Add local section database and image database
添加本地章节数据库，图片数据库。离线可以加载。解析完成的不会再解析失败。

+++ app/src/main/java/com/hiroshi/cimoc/ui/activity/ReaderActivity.java
@@ -409,12 +409,12 @@ public abstract class ReaderActivity extends BaseActivity implements OnTapGestur
     }
 
     @Override
-    public void onImageLoadSuccess(int id, String url) {
+    public void onImageLoadSuccess(Long id, String url) {
         mReaderAdapter.update(id, url);
     }
 
     @Override
-    public void onImageLoadFail(int id) {
+    public void onImageLoadFail(Long id) {
         mReaderAdapter.update(id, null);
     }
